split services and utils
