+++ src/app/auth/components/Register.vue
@@ -113,7 +113,7 @@
 </template>
 
 <script>
-import { rules } from '../../shared/services/validators';
+import { rules } from '../../utils/validators';
 import { registerSuccess } from '../+store/auth-state';
 import { mapActions } from 'vuex';
