--- src/app/shared/services/index.js
@@ -1,5 +1,0 @@
-export * from './httpClient';
-export * from './firebaseConfig';
-export * from './validators';
-export * from './guards';
-export * from './toasted';
